+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## vNext
+
+- Do not fail if keyring is not accessible
+  ([issues/45](https://github.com/vlaci/openconnect-sso/issues/45),
+  [pull/46](https://github.com/vlaci/openconnect-sso/pull/46))
+
 ## v0.7.0
 
 - It is now possible to reuse previous authentication sessions, so that